Updated readme, updated names and some extra functionality for C#
