+++ sdks/client-csharp/FeatureHubTestProject/ConfigTest.cs
@@ -9,11 +9,11 @@ namespace FeatureHubTestProject
     [Test]
     public void EnsureConfigCorrectlyDeterminesUrl()
     {
-      var cfg = new FeatureHubConfig("http://localhost:80/", "123*123");
+      var cfg = new EdgeFeatureHubConfig("http://localhost:80/", "123*123");
       Assert.IsTrue(!cfg.ServerEvaluation);
       Assert.AreEqual("http://localhost:80/features/123*123", cfg.Url);
 
-      cfg = new FeatureHubConfig("http://localhost:80", "123");
+      cfg = new EdgeFeatureHubConfig("http://localhost:80", "123");
       Assert.IsTrue(cfg.ServerEvaluation);
       Assert.AreEqual("http://localhost:80/features/123", cfg.Url);
     }
